+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/refs/SmkRuleOrCheckpointNameReference.kt
@@ -1,8 +1,9 @@
 package com.jetbrains.snakecharm.lang.psi.impl.refs
 
 import com.intellij.lang.annotation.HighlightSeverity
-import com.intellij.psi.util.elementType
-import com.intellij.psi.util.parentOfType
+import com.intellij.openapi.module.ModuleUtilCore
+import com.intellij.psi.search.GlobalSearchScope
+import com.intellij.psi.stubs.StubIndex
 import com.jetbrains.python.psi.AccessDirection
 import com.jetbrains.python.psi.impl.references.PyReferenceImpl
 import com.jetbrains.python.psi.resolve.PyResolveContext
@@ -11,13 +12,11 @@ import com.jetbrains.python.psi.types.TypeEvalContext
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.codeInsight.resolve.SmkResolveUtil
 import com.jetbrains.snakecharm.lang.psi.SmkFile
+import com.jetbrains.snakecharm.lang.psi.SmkModule
 import com.jetbrains.snakecharm.lang.psi.SmkReferenceExpression
-import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpoint
-import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkElementTypes
-import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkStubElementTypes
+import com.jetbrains.snakecharm.lang.psi.stubs.SmkModuleNameIndex
 import com.jetbrains.snakecharm.lang.psi.types.SmkCheckpointType
 import com.jetbrains.snakecharm.lang.psi.types.SmkRulesType
-import com.jetbrains.snakecharm.lang.psi.types.SmkUsesType
 
 class SmkRuleOrCheckpointNameReference(
     element: SmkReferenceExpression,
@@ -50,53 +49,36 @@ class SmkRuleOrCheckpointNameReference(
 
         results.addAll(SmkRulesType(null, smkFile).resolveMember(name, element, ctx, myContext))
         results.addAll(SmkCheckpointType(null, smkFile).resolveMember(name, element, ctx, myContext))
-        results.addAll(SmkUsesType(null, smkFile).resolveMember(name, element, ctx, myContext))
-        results.addAll(collectModulesAndResolveThem(smkFile, name))
-        results.addAll(collectModuleFromUseSection(element))
+        results.addAll(collectModulesAndResolveThem(smkFile, element))
 
         return results
     }
 
     /**
-     * Collects all modules sections names from local file which name is [name]
+     * Collects all modules sections names from local file or using indexes which name is [element] name
      */
-    private fun collectModulesAndResolveThem(smkFile: SmkFile, name: String): List<RatedResolveResult> {
-        val modules = smkFile.collectModules().map { it.second }.filter { elem -> elem.name == name }
+    private fun collectModulesAndResolveThem(
+        smkFile: SmkFile,
+        element: SmkReferenceExpression
+    ): List<RatedResolveResult> {
+        val module = element.let { ModuleUtilCore.findModuleForPsiElement(it.originalElement) }
+        val target = element.referencedName ?: return emptyList()
+        val modules = if (module == null) {
+            smkFile.collectModules().map { it.second }.filter { modulePsi -> modulePsi.name == target }
+        } else {
+            StubIndex.getElements(
+                SmkModuleNameIndex.KEY,
+                target,
+                module.project,
+                GlobalSearchScope.moduleWithDependentsScope(module),
+                SmkModule::class.java
+            )
+        }
         if (modules.isEmpty()) {
             return emptyList()
         }
-        return modules.map { element ->
-            RatedResolveResult(SmkResolveUtil.RATE_NORMAL, element)
+        return modules.map { variant ->
+            RatedResolveResult(SmkResolveUtil.RATE_NORMAL, variant)
         }
-    }
-
-    /**
-     * Resolve rule reference, which is declared in 'use' section.
-     * It refers to module, which imports such rule.
-     * If there no such module, return empty array.
-     */
-    private fun collectModuleFromUseSection(
-        element: SmkReferenceExpression
-    ): List<RatedResolveResult> {
-        if (element.parent.elementType != SmkElementTypes.USE_IMPORTED_RULES_NAMES) {
-            return emptyList()
-        }
-        val parent = element.parentOfType<SmkRuleOrCheckpoint>()
-        if (parent != null && parent.elementType == SmkStubElementTypes.USE_DECLARATION_STATEMENT) {
-            var moduleRef = element.parent.nextSibling
-            while (moduleRef != null && moduleRef.elementType != SmkElementTypes.REFERENCE_EXPRESSION) {
-                moduleRef = moduleRef.nextSibling
-            }
-            if (moduleRef != null) {
-                return listOf(
-                    RatedResolveResult(
-                        SmkResolveUtil.RATE_NORMAL,
-                        (moduleRef as SmkReferenceExpression).reference.resolve()
-                    )
-                )
-            }
-        }
-
-        return emptyList()
     }
 }